test1/requirements.txt
Flask==1.1.2
line-bot-sdk==1.16.0
psycopg2==2.8.1
requests==2.24.0